requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.3.0
marshmallow-sqlalchemy==1.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
